Add README.md and factor out products per page filter functionality in ProductsPerPageFilter component
