+++ src/components/ProductsPerPageFilter/ProductsPerPageFilter.scss
@@ -1,0 +1,8 @@
+.products-per-page select {
+  padding: 5px 8px;
+  margin-left: 10px;
+  border-radius: 5px;
+  border: none;
+  cursor: pointer;
+  background-color: var(--secondary-ultralight-color);
+}
