+++ src/components/ProductsPerPageFilter/ProductsPerPageFilter.tsx
@@ -1,0 +1,32 @@
+import { ChangeEvent } from "react";
+import "./ProductsPerPageFilter.scss";
+
+type ProductsPerPageFilterProps = {
+  setProductsPerPage: (num: number) => void;
+};
+
+export const ProductsPerPageFilter = ({
+  setProductsPerPage,
+}: ProductsPerPageFilterProps): React.ReactNode => {
+  const handleProductsAmount = (event: ChangeEvent<HTMLSelectElement>) => {
+    setProductsPerPage(+event.target.value);
+  };
+
+  return (
+    <div className="products-per-page">
+      <label htmlFor="products-amount">Products per page: </label>
+      <select
+        name="products-amount"
+        id="products-amount"
+        onChange={handleProductsAmount}
+        defaultValue={10}
+      >
+        <option value="5">5</option>
+        <option value="10">10</option>
+        <option value="15">15</option>
+        <option value="25">25</option>
+        <option value="50">50</option>
+      </select>
+    </div>
+  );
+};
