+++ src/components/index.ts
@@ -6,3 +6,4 @@ export * from './Pagination';
 export * from './ProductForm';
 export * from './ProductItem';
 export * from './ProductList';
+export * from './ProductsPerPageFilter';
